virtio-snd: move uio_sound.h to tools
